Committing manually all changes after long auto run + manual fixes
date for the commit is chosen based on the latest dates in
.dandi/assets.json

various deeds were done manually to bring it to some sane state, e.g.
to sync some zarrs which were found not to be synced etc.  It might
incomplete, but FTR here is an example command which was used to
populate some zarrs which were still missing but listed in
dandi/assets.json:

  grep '\.zarr' .dandi/assets.json | sed -e 's/[",]//g' | while read zarr; do grep -q "$zarr" .gitmodules || echo "$zarr"; done | xargs python ../tools/get_zarr_ids_by_path.py 000108 | while read path id; do python ../tools/sync_zarrs_by_id.py 000108 $id; datalad -C ../../dandizarrs/$id push --to=github; datalad clone -d .  https://github.com/dandizarrs/$id $path; datalad uninstall $path; done

it seems that

  jq -r '.[].path' < .dandi/assets.json | while read p; do /bin/ls -d "$p" >/dev/null || echo $p; done

came up empty so we have something for every path listed in
.dandi/assets.json